pyproject.toml
[project]
name = "dytalabs-io-install-secure-archlinux"
version = "0.1.0"
description = ""
authors = [
    {name = "Your Name",email = "you@example.com"}
]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "ansible-core (>=2.18.2,<3.0.0)",
    "ansible-compat (>=25.1.2,<26.0.0)",
    "askpass (>=0.1,<0.2)",
    "dnspython (>=2.7.0,<3.0.0)",
    "jinja2 (>=3.1.5,<4.0.0)",
    "jinja2-ansible-filters (>=1.3.2,<2.0.0)",
    "jmespath (>=1.0.1,<2.0.0)",
    "netaddr (>=1.3.0,<2.0.0)",
    "passlib (>=1.7.4,<2.0.0)",
    "pywinrm (>=0.5.0,<0.6.0)"
]

[tool.poetry]
packages = [{include = "dytalabs_io"}]


[tool.poetry.group.dev.dependencies]
ansible-lint = "^25.1.2"
ansible-runner = "^2.4.0"
ansible-bender = "^0.10.1"
molecule = "^25.2.0"
molecule-vagrant = "^2.0.0"
autopep8 = "^2.3.2"
black = "^25.1.0"
build = "^1.2.2.post1"
flake8 = "^7.1.1"
installer = "^0.7.0"
mypy = "^1.15.0"
pylint = "^3.3.4"
pytest = "^8.3.4"
pytest-testinfra = "^10.1.1"
pytest-xdist = "^3.6.1"
setuptools = "^75.8.0"
six = "^1.17.0"
wheel = "^0.45.1"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
